backend/requirements.txt
pytest==7.4.4
pytest-django==4.8.0
